refactor(server): organize context tree services
