+++ packages/server/src/__tests__/context-reviewer-pr.test.ts
@@ -21,7 +21,7 @@ import {
   handleContextReviewerPullRequest as handleContextReviewerPullRequestService,
   normalizeGithubRepo,
   renderContextReviewerPrPrompt,
-} from "../services/context-reviewer-pr.js";
+} from "../services/context-tree/reviewer/github/pr.js";
 import { createMember } from "../services/member.js";
 import { putOrgSetting } from "../services/org-settings.js";
 import { upsertInstallationFromMetadata } from "../services/scm/github/app-installations.js";
