+++ packages/server/src/__tests__/context-reviewer-publisher.test.ts
@@ -12,8 +12,8 @@ import { messages } from "../db/schema/messages.js";
 import { organizations } from "../db/schema/organizations.js";
 import { createAgent } from "../services/agent.js";
 import { bindAgentRuntimeSession } from "../services/agent-runtime-session.js";
-import { handleContextReviewerPrEvent as handleContextReviewerPrEventService } from "../services/context-reviewer-pr.js";
-import { submitContextReviewOutcome } from "../services/context-reviewer-publisher.js";
+import { handleContextReviewerPrEvent as handleContextReviewerPrEventService } from "../services/context-tree/reviewer/github/pr.js";
+import { submitContextReviewOutcome } from "../services/context-tree/reviewer/github/publisher.js";
 import { putOrgSetting } from "../services/org-settings.js";
 import { upsertInstallationFromMetadata } from "../services/scm/github/app-installations.js";
 import { createAdminContext, seedHealthyAgentRuntime, useTestApp } from "./helpers.js";
