+++ packages/server/src/__tests__/github-app-token.test.ts
@@ -1,7 +1,7 @@
 import { generateKeyPairSync } from "node:crypto";
 import type { ContextTreeSnapshot } from "@first-tree/shared";
 import { beforeAll, describe, expect, it } from "vitest";
-import type { ContextTreeBinding } from "../services/context-tree-snapshot.js";
+import type { ContextTreeBinding } from "../services/context-tree/snapshot.js";
 import type { InstallationRow } from "../services/scm/github/app-installations.js";
 import {
   type ContextTreeInstallationTokenResult,
