+++ packages/server/src/__tests__/service-branch-defaults.test.ts
@@ -16,7 +16,7 @@ import {
 } from "../services/agent.js";
 import { assertNoRuntimeSwitchInProgress, getRuntimeSwitchClaim } from "../services/agent-runtime-switch.js";
 import { createChat, resolveAgentIdsByNameInOrg, updateChatMetadata } from "../services/chat.js";
-import { explainContextTreeIoDecision } from "../services/context-tree-io.js";
+import { explainContextTreeIoDecision } from "../services/context-tree/io.js";
 import { maybeUnwrapDoubleEncoded, preflightMessageSendIntent } from "../services/message.js";
 import { createResourcesService } from "../services/resources.js";
 import {
