+++ packages/server/src/api/webhooks/gitlab.ts
@@ -2,7 +2,7 @@ import type { FastifyInstance } from "fastify";
 import { BadRequestError, NotFoundError } from "../../errors.js";
 import { createLogger } from "../../observability/index.js";
 import { invalidateChatAudience } from "../../services/chat-audience-cache.js";
-import { handleContextReviewerMrEvent } from "../../services/context-reviewer-mr.js";
+import { handleContextReviewerMrEvent } from "../../services/context-tree/reviewer/gitlab/mr.js";
 import { runDeferredSendMessagePostCommitEffects } from "../../services/message.js";
 import { notifyRecipients } from "../../services/notifier.js";
 import {
