+++ packages/server/src/services/context-tree/activation.ts
@@ -5,11 +5,11 @@ import {
   resolveGitLabRepositoryWebIdentity,
 } from "@first-tree/shared";
 import { eq } from "drizzle-orm";
-import type { Database } from "../db/connection.js";
-import { gitlabConnections } from "../db/schema/gitlab-connections.js";
-import type { OrgScope } from "../scope/types.js";
-import { getOrgContextTreeSettingState } from "./org-settings.js";
-import { getOrganization } from "./organization.js";
+import type { Database } from "../../db/connection.js";
+import { gitlabConnections } from "../../db/schema/gitlab-connections.js";
+import type { OrgScope } from "../../scope/types.js";
+import { getOrgContextTreeSettingState } from "../org-settings.js";
+import { getOrganization } from "../organization.js";
 
 const CONTEXT_TREE_SETTINGS_URL = "/settings/context#binding";
