+++ packages/server/src/services/context-tree/io.ts
@@ -20,16 +20,16 @@ import {
   toolFileRefSchema,
 } from "@first-tree/shared";
 import { and, desc, eq, gte, inArray, lt, lte, or, type SQL, sql } from "drizzle-orm";
-import type { Database } from "../db/connection.js";
-import { agents } from "../db/schema/agents.js";
-import { chatMembership } from "../db/schema/chat-membership.js";
-import { chats } from "../db/schema/chats.js";
-import { contextTreeIoEvents } from "../db/schema/context-tree-io-events.js";
-import { sessionEvents } from "../db/schema/session-events.js";
-import { BadRequestError } from "../errors.js";
-import { createLogger } from "../observability/index.js";
-import { type TimingSink, timeSyncWithSink, timeWithSink } from "../observability/timing.js";
-import { getOrgContextTreeBinding } from "./org-settings.js";
+import type { Database } from "../../db/connection.js";
+import { agents } from "../../db/schema/agents.js";
+import { chatMembership } from "../../db/schema/chat-membership.js";
+import { chats } from "../../db/schema/chats.js";
+import { contextTreeIoEvents } from "../../db/schema/context-tree-io-events.js";
+import { sessionEvents } from "../../db/schema/session-events.js";
+import { BadRequestError } from "../../errors.js";
+import { createLogger } from "../../observability/index.js";
+import { type TimingSink, timeSyncWithSink, timeWithSink } from "../../observability/timing.js";
+import { getOrgContextTreeBinding } from "../org-settings.js";
 
 const CONTEXT_TREE_IO_FEED_LIMIT = 50;
 // Grep/Glob count as reads at the granularity their refs carry: the client
