+++ packages/server/src/services/context-tree/repo-provisioner.ts
@@ -4,10 +4,10 @@ import {
   GithubAppApiError,
   getRepository,
   verifyUserCanAdministerInstallation,
-} from "./scm/github/app.js";
-import type { InstallationRow } from "./scm/github/app-installations.js";
-import { createUserRepo, GithubApiError, type GithubCreatedRepo } from "./scm/github/oauth.js";
-import type { GithubUserToken } from "./scm/github/user-token.js";
+} from "../scm/github/app.js";
+import type { InstallationRow } from "../scm/github/app-installations.js";
+import { createUserRepo, GithubApiError, type GithubCreatedRepo } from "../scm/github/oauth.js";
+import type { GithubUserToken } from "../scm/github/user-token.js";
 
 /**
  * Owner-agnostic Context Tree repo provisioner.
