+++ packages/server/src/services/context-tree/reviewer/settings.ts
@@ -12,23 +12,23 @@ import {
   type SetupBlocker,
 } from "@first-tree/shared";
 import { and, eq, sql } from "drizzle-orm";
-import type { Database } from "../db/connection.js";
-import { agents } from "../db/schema/agents.js";
-import { clients } from "../db/schema/clients.js";
-import { githubAppInstallations } from "../db/schema/github-app-installations.js";
-import { gitlabConnections } from "../db/schema/gitlab-connections.js";
-import { members } from "../db/schema/members.js";
-import { organizationSettings } from "../db/schema/organization-settings.js";
-import { organizations } from "../db/schema/organizations.js";
-import { ConflictError, NotFoundError } from "../errors.js";
-import { agentNotLandingCampaignTrialCondition } from "./access-control.js";
-import { getOrgContextReviewRuntime, getOrgSetting, isOrgContextReviewRuntimeCurrent } from "./org-settings.js";
-import { findInstallationByOrg } from "./scm/github/app-installations.js";
+import type { Database } from "../../../db/connection.js";
+import { agents } from "../../../db/schema/agents.js";
+import { clients } from "../../../db/schema/clients.js";
+import { githubAppInstallations } from "../../../db/schema/github-app-installations.js";
+import { gitlabConnections } from "../../../db/schema/gitlab-connections.js";
+import { members } from "../../../db/schema/members.js";
+import { organizationSettings } from "../../../db/schema/organization-settings.js";
+import { organizations } from "../../../db/schema/organizations.js";
+import { ConflictError, NotFoundError } from "../../../errors.js";
+import { agentNotLandingCampaignTrialCondition } from "../../access-control.js";
+import { getOrgContextReviewRuntime, getOrgSetting, isOrgContextReviewRuntimeCurrent } from "../../org-settings.js";
+import { findInstallationByOrg } from "../../scm/github/app-installations.js";
 import {
   type GithubReviewCredentials,
   type GithubReviewProbeResult,
   getTeamSetupCapabilities,
-} from "./setup-capabilities.js";
+} from "../../setup-capabilities.js";
 
 export type ContextReviewerMutationOptions = {
   updatedBy: string;
