+++ packages/server/src/services/setup-capabilities.ts
@@ -12,7 +12,7 @@ import {
 import { eq } from "drizzle-orm";
 import type { Database } from "../db/connection.js";
 import { gitlabConnections } from "../db/schema/gitlab-connections.js";
-import { readContextReviewerAgentReadiness } from "./context-reviewer-readiness.js";
+import { readContextReviewerAgentReadiness } from "./context-tree/reviewer/readiness.js";
 import { getOrgContextReviewRuntime } from "./org-settings.js";
 import {
   createAppJwt,
