+++ packages/server/src/services/team-agent-settings.ts
@@ -10,8 +10,8 @@ import type { Database } from "../db/connection.js";
 import { organizationSettings } from "../db/schema/organization-settings.js";
 import { organizations } from "../db/schema/organizations.js";
 import { ConflictError, NotFoundError } from "../errors.js";
-import { loadValidContextReviewerAgent } from "./context-reviewer-common.js";
-import { listContextReviewerCandidates } from "./context-reviewer-readiness.js";
+import { loadValidContextReviewerAgent } from "./context-tree/reviewer/common.js";
+import { listContextReviewerCandidates } from "./context-tree/reviewer/readiness.js";
 import { getOrgSetting } from "./org-settings.js";
 import { findInstallationByOrg, type InstallationRow } from "./scm/github/app-installations.js";
